fix Integer cannot be cast to class Long

+++ core/src/main/java/com/ctrip/framework/drc/core/driver/binlog/impl/AbstractRowsEvent.java
@@ -376,7 +376,7 @@ public abstract class AbstractRowsEvent extends AbstractLogEvent implements Rows
                 if (column.isUnsigned()) {
                     ByteHelper.writeUnsignedIntLittleEndian((long) value, out);
                 } else {
-                    ByteHelper.writeIntLittleEndian((long) value, out);
+                    ByteHelper.writeIntLittleEndian((int) value, out);
                 }
                 return;
